Auto-format YAML files
There are a few minor formatting differences between Prettier and
yamlfix. The changes made by yamlfix seem to be compatible with Prettier
so the files should be synced up going forward.

An override has been added for Prettier to force the use of single
quotes in YAML files. This was the only setting that was causing
problems when running the pre-commit hook.