Add API session and REST functions
The api module contains a global API session which is used when making
calls to the CWMS Data API. The user may specify an alternative root URL
for the API and an optional authentication key when initializing the
session.

General REST functions (get, post, etc) are provided for interacting
with the CDA via the session object.